Update synsaverestore.sp
Added cvar sm_disabletransition default 0. This will disable all transition saves. The ones that are required are rebuilt, for example, d1_town_03 to d1_town_02 transition will not transition any entities or player data, but you will spawn in the correct transition spawn.
This has also been done for d2_coast_08 to 07, and ep2_outland_04 to 02.